requirements.txt
altair
numpy
pandas
pydeck
streamlit
azure.identity
langchain
openai
langchain_openai
fpdf
os